# Decrease buffer size to 128KiB to improve performance when reading from pipes.